feat: update Testcontainers packages to version 4.10.0 and SSH.NET to version 2025.1.0

+++ src/Core/TestcontainersDockerManager.cs
@@ -251,7 +251,7 @@ public class TestcontainersDockerManager : IDockerContainerManager
             var fileBytes = await File.ReadAllBytesAsync(context.Source);
             const UnixFileModes fileMode = UnixFileModes.UserRead | UnixFileModes.UserWrite | 
                                            UnixFileModes.GroupRead | UnixFileModes.OtherRead;
-            await _container.CopyAsync(fileBytes, context.Destination, fileMode);
+            await _container.CopyAsync(fileBytes, context.Destination, (uint)fileMode);
         }
         catch (Exception ex)
         {
